delete InMemoryTweetRepository.ts
