+++ src/application/TweetApplicationService.ts
@@ -1,9 +1,5 @@
 import Tweet from '../domain/models/Tweet/ConcreteClasses/Tweet';
-import {
-  ITweetRepository,
-  TweetCreateProps,
-} from '../domain/models/Tweet/ITweetRepository';
-import { InMemoryTweetRepository } from '../inMemory/InMemoryTweetRepository';
+import { TweetCreateProps } from '../domain/models/Tweet/ITweetRepository';
 import { AbstractTweet } from '../domain/models/Tweet/AbstractTweet';
 import { TempTweetData } from '../infrastructure/TempTweetDataModel';
 import { TweetDataModel } from '../infrastructure/TweetDataModel';
@@ -11,10 +7,7 @@ import { TweetRepository } from '../infrastructure/TweetRepository';
 import { TweetFactory } from '../domain/models/Tweet/TweetFactory';
 
 export class TweetApplicationService {
-  static readonly tweetRepository: ITweetRepository = new InMemoryTweetRepository();
-
-  // todo tweetRepo を1つにまとめよう
-  static readonly tweetRepositoryNew = new TweetRepository();
+  static readonly tweetRepository = new TweetRepository();
 
   static readonly tweetFactory = new TweetFactory();
 
@@ -40,11 +33,11 @@ export class TweetApplicationService {
   }
 
   static fetchTimeline(): Promise<Response> {
-    return TweetApplicationService.tweetRepositoryNew.fetchTimeline();
+    return TweetApplicationService.tweetRepository.fetchTimeline();
   }
 
   static postTweet(tweetDataModel: TempTweetData): Promise<Response> {
-    return TweetApplicationService.tweetRepositoryNew.postTweet(tweetDataModel);
+    return TweetApplicationService.tweetRepository.postTweet(tweetDataModel);
   }
 
   static returnTweetId(tweet: Tweet): string {
